pyproject.toml
[project]
name = "bi-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chardet>=5.2.0",
    "fastapi>=0.128.0",
    "jupyter-client>=8.6.3",
    "langchain>=1.2.6",
    "langchain-anthropic>=1.3.4",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.6",
    "lightgbm>=4.6.0",
    "numpy>=2.4.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "prophet>=1.3.0",
    "pyarrow>=22.0.0",
    "pydantic>=2.12.5",
    "python-calamine>=0.6.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "scikit-learn>=1.8.0",
    "setuptools>=82.0.0",
    "statsmodels>=0.14.6",
    "uvicorn>=0.40.0",
    "xlrd>=2.0.2",
]
